requirements.txt
psycopg2-binary==2.9.10
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
beautifulsoup4
lxml
fastapi
uvicorn[standard]